requirements.txt
# Core library dependency
PyNaCl>=1.5.0
